muuta hookin ja sen palauttaman funktion nimeä ja laita argumentit object-muotoon
